Added boolean powerUpUsed in Movements so when player jumps, it is only used once

+++ core/src/com/logic/strategy/Context.java
@@ -7,7 +7,7 @@ public class Context {
         this.strategy = strategy;
     }
 
-    public int executeStrategy(int value) {
+    public float executeStrategy(float value) {
         return strategy.doOperation(value);
     }
 }
