+++ core/src/com/logic/strategy/MoveFaster.java
@@ -3,7 +3,7 @@ package com.logic.strategy;
 public class MoveFaster implements Strategy {
 
     @Override
-    public int doOperation(int value) {
+    public float doOperation(float value) {
         return value*2;
     }
 }
